pyproject.toml
[project]
name = "asteroid-belt"
version = "0.1.0"
description = "DLMM strategy research desk for Meteora pools"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.7",
    "polars>=1.0",
    "duckdb>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "openai>=1.50",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-pyyaml>=6.0",
    "respx>=0.21",
]

[project.scripts]
belt = "asteroid_belt.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["asteroid_belt"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
